requirements.txt
# FastAPI dependencies
fastapi>=0.95.0
uvicorn>=0.21.1
pydantic>=1.10.7
python-dotenv>=1.0.0
 
# Database
sqlalchemy>=2.0.9
alembic>=1.10.3
 
# Streamlit
streamlit>=1.21.0
pandas>=2.0.0
numpy>=1.24.2
 
# HTTP client for API calls
requests>=2.28.2
 
# Testing
pytest>=7.3.1

altair==5.5.0
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
autogen-agentchat==0.7.1
autogen-core==0.7.1
blinker==1.9.0
cachetools==6.1.0
certifi==2025.8.3
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
distro==1.9.0
docx==0.2.4
fastapi==0.116.1
gitdb==4.0.12
GitPython==3.1.45
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib_metadata==8.7.0
iniconfig==2.1.0
Jinja2==3.1.6
jiter==0.10.0
jsonref==1.1.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
lxml==6.0.0
Mako==1.3.10
MarkupSafe==3.0.2
narwhals==2.0.1
openai==1.98.0
opentelemetry-api==1.36.0
packaging==25.0
pillow==11.3.0
pluggy==1.6.0
protobuf==5.29.5
pyarrow==21.0.0
pyautogen==0.10.0
pydeck==0.9.1
Pygments==2.19.2
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
pytz==2025.2
referencing==0.36.2
rpds-py==0.26.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
starlette==0.47.2
tenacity==9.1.2
toml==0.10.2
tornado==6.5.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
watchdog==6.0.0
zipp==3.23.0
autogen_ext
tiktoken
python-multipart